+++ gral-core/src/test/java/de/erichseifert/gral/io/plots/PlotsTests.java
@@ -27,7 +27,8 @@ import org.junit.runners.Suite;
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
 	DrawableWriterFactoryTest.class,
-	DrawableWriterTest.class
+	DrawableWriterTest.class,
+	VectorExportTest.class
 })
 public class PlotsTests {
 }